Fire and Smoke Detector in Smart House
To create a smart home by using IoT devises such as a fire and smoke detector that connect directly to the fire station to reduce wasted time by using a smartphone and two Ip phones that can call each other. the smart home can control the damages until the fire people come to the rescue, also showing a warring text and an image to anyone gets close to it.